pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wikipedia-rag-system"
version = "0.1.0"
description = "Wikipedia-based RAG system with local LLM on AWS"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Riku Miura", email = "riku-miura@example.com"}
]

dependencies = [
    "langchain>=0.1.0",
    "sentence-transformers>=2.2.2",
    "faiss-cpu>=1.7.4",
    "wikipedia-api>=0.6.0",
    "boto3>=1.34.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.1.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "moto[all]>=5.0.0",
    "hypothesis>=6.92.0",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | cdk.out
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "backend/tests",
]
pythonpath = [
    "backend/src"
]
